Use open4 and timeout to avoid hanging processes

+++ test/test_osxscreenshot.rb
@@ -2,7 +2,7 @@ require "test/unit"
 require "osxscreenshot"
 
 class TestOsxscreenshot < Test::Unit::TestCase
-  
+
   def test_loads_url
     @tmpfile = OSX::Screenshot.capture("http://example.com")
     assert_not_nil @tmpfile
@@ -15,15 +15,25 @@ class TestOsxscreenshot < Test::Unit::TestCase
     extra_file_name = @tmpfile.gsub(/-0\.png$/, "-1.png")
     assert !File.exist?(extra_file_name)
   end
-  
+
   def test_uses_custom_tmpdir
     @tmpfile = OSX::Screenshot.capture("http://example.com", :tmpdir => "./tmp")
     assert_match(/^\.\/tmp/, @tmpfile)
     assert File.exist?(@tmpfile)
   end
-  
+
+  def test_handles_timeout
+    @tmpfile = OSX::Screenshot.capture("http://example.com", {
+                                         :timeout => 1,
+                                         :webkit2png => "sleep 5 &&"
+                                       })
+    assert_nil @tmpfile
+  end
+
   def teardown
-    FileUtils.rm @tmpfile
+    if @tmpfile
+      FileUtils.rm @tmpfile
+    end
   end
 
 end
